Extract refresh handling in to its own module.
Sets the groundwork for doing dispatch based on device type
instead of needing a ton of `if`'s everywhere.
