+++ drivers/SmartThings/philips-hue/src/hue_device_types.lua
@@ -1,0 +1,4 @@
+return {
+  BRIDGE = "bridge",
+  LIGHT = "light"
+}
